agent/requirements.txt
openai-agents
fastapi
uvicorn[standard]
httpx
python-dotenv
